itest: move multi hop tests to own files
PURE CODE MOVE.
